requirements.txt
streamlit
pdfplumber
python-dotenv
pandas
google-generativeai
